backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain and LangGraph
langchain==0.1.6
langgraph==0.0.20
langchain-google-genai==0.0.6
langchain-community==0.0.20

# AI/ML Libraries
google-generativeai==0.3.2
ollama==0.1.6
chromadb==0.4.22
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0

# Config and Parsing
pyyaml==6.0.1
lxml==5.1.0
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
